Move server start into mod
